Implement a position measurements (gauge) class
