+++ examples/dtqw_1d2p_profiling.py
@@ -5,6 +5,7 @@ import logging
 from pyspark import SparkContext, SparkConf
 
 from sparkquantum.dtqw.coin.coin1d.hadamard1d import Hadamard1D
+from sparkquantum.dtqw.gauge.position_gauge import PositionGauge
 from sparkquantum.dtqw.interaction.collision_phase_interaction import CollisionPhaseInteraction
 from sparkquantum.dtqw.mesh.mesh1d.line import Line
 from sparkquantum.dtqw.state import State
@@ -143,7 +144,11 @@ final_state = dtqw.walk(steps, initial_state)
 final_state.profiler = profiler
 
 # Measuring the state of the system and plotting its PDF
-joint, collision, marginal = final_state.measure()
+gauge = PositionGauge()
+
+gauge.profiler = profiler
+
+joint, collision, marginal = gauge.measure(final_state)
 joint.plot(walk_path + 'joint_1d2p', dpi=300)
 joint.plot_contour(walk_path + 'joint_1d2p_contour', dpi=300)
 collision.plot(walk_path + 'collision_1d2p', dpi=300)
