+++ examples/dtqw_2d1p_profiling.py
@@ -4,6 +4,7 @@ import logging
 from pyspark import SparkContext, SparkConf
 
 from sparkquantum.dtqw.coin.coin2d.hadamard2d import Hadamard2D
+from sparkquantum.dtqw.gauge.position_gauge import PositionGauge
 from sparkquantum.dtqw.mesh.mesh2d.diagonal.lattice import LatticeDiagonal
 from sparkquantum.dtqw.state import State
 from sparkquantum.dtqw.qw_profiler import QuantumWalkProfiler
@@ -102,7 +103,11 @@ final_state = dtqw.walk(steps, initial_state)
 final_state.profiler = profiler
 
 # Measuring the state of the system and plotting its PDF
-joint = final_state.measure()
+gauge = PositionGauge()
+
+gauge.profiler = profiler
+
+joint = gauge.measure(final_state)
 joint.plot(walk_path + 'joint_2d1p', dpi=300)
 joint.plot_contour(walk_path + 'joint_2d1p_contour', dpi=300)
